Organize gqlgen files
