+++ backend/graph/following_teacher.graphqls
@@ -1,0 +1,25 @@
+# GraphQL schema example
+#
+# https://gqlgen.com/getting-started/
+
+type Teacher {
+  id: ID!
+  name: String!
+}
+
+extend type Query {
+  followingTeachers: [Teacher!]!
+}
+
+#type Query {
+#  todos: [Todo!]!
+#}
+#
+#input NewTodo {
+#  text: String!
+#  userId: String!
+#}
+#
+#type Mutation {
+#  createTodo(input: NewTodo!): Todo!
+#}
