+++ backend/graph/following_teacher.resolvers.go
@@ -1,0 +1,15 @@
+package graph
+
+// This file will be automatically regenerated based on the schema, any resolver implementations
+// will be copied through when generating and any unknown code will be moved to the end.
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/oinume/lekcije/backend/graph/model"
+)
+
+func (r *queryResolver) FollowingTeachers(ctx context.Context) ([]*model.Teacher, error) {
+	panic(fmt.Errorf("not implemented"))
+}
