+++ backend/graph/resolver/resolver.go
@@ -1,4 +1,4 @@
-package graph
+package resolver
 
 // This file will not be regenerated automatically.
 //
